refactor(layout): group the layout components in a FinchAppLayout folder
