+++ src/index.ts
@@ -28,33 +28,33 @@ export type { TiledHeatmapSelectorProps } from './features/TiledHeatmapSelector'
 export { default as Beamstop } from './features/Beamstop';
 export type { BeamstopProps } from './features/Beamstop';
 
-export { default as FinchAppLayout } from './components/FinchAppLayout';
-export type { FinchAppLayoutProps } from './components/FinchAppLayout';
+export { default as FinchAppLayout } from './components/FinchAppLayout/FinchAppLayout';
+export type { FinchAppLayoutProps } from './components/FinchAppLayout/FinchAppLayout';
 
-export { default as FinchHeader } from './components/FinchHeader';
-export type { FinchHeaderProps } from './components/FinchHeader';
+export { default as FinchHeader } from './components/FinchAppLayout/FinchHeader';
+export type { FinchHeaderProps } from './components/FinchAppLayout/FinchHeader';
 
-export { default as FinchMainContent } from './components/FinchMainContent';
-export type { FinchMainContentProps } from './components/FinchMainContent';
+export { default as FinchMainContent } from './components/FinchAppLayout/FinchMainContent';
+export type { FinchMainContentProps } from './components/FinchAppLayout/FinchMainContent';
 
-export { default as FinchPageTabs } from './components/FinchPageTabs';
-export type { FinchPageTabsProps } from './components/FinchPageTabs';
+export { default as FinchPageTabs } from './components/FinchAppLayout/FinchPageTabs';
+export type { FinchPageTabsProps } from './components/FinchAppLayout/FinchPageTabs';
 
-export { default as FinchSidebar } from './components/FinchSidebar';
-export type { FinchSidebarProps } from './components/FinchSidebar';
+export { default as FinchSidebar } from './components/FinchAppLayout/FinchSidebar';
+export type { FinchSidebarProps } from './components/FinchAppLayout/FinchSidebar';
 
 // @deprecated — use Finch* equivalents
-export { default as HubAppLayout } from './components/FinchAppLayout';
-export type { FinchAppLayoutProps as HubAppLayoutProps } from './components/FinchAppLayout';
+export { default as HubAppLayout } from './components/FinchAppLayout/FinchAppLayout';
+export type { FinchAppLayoutProps as HubAppLayoutProps } from './components/FinchAppLayout/FinchAppLayout';
 
-export { default as HubHeader } from './components/FinchHeader';
-export type { FinchHeaderProps as HubHeaderProps } from './components/FinchHeader';
+export { default as HubHeader } from './components/FinchAppLayout/FinchHeader';
+export type { FinchHeaderProps as HubHeaderProps } from './components/FinchAppLayout/FinchHeader';
 
-export { default as HubMainContent } from './components/FinchMainContent';
-export type { FinchMainContentProps as HubMainContentProps } from './components/FinchMainContent';
+export { default as HubMainContent } from './components/FinchAppLayout/FinchMainContent';
+export type { FinchMainContentProps as HubMainContentProps } from './components/FinchAppLayout/FinchMainContent';
 
-export { default as HubSidebar } from './components/FinchSidebar';
-export type { FinchSidebarProps as HubSidebarProps } from './components/FinchSidebar';
+export { default as HubSidebar } from './components/FinchAppLayout/FinchSidebar';
+export type { FinchSidebarProps as HubSidebarProps } from './components/FinchAppLayout/FinchSidebar';
 
 export { default as ContainerQServer } from './components/QServer/ContainerQServer';
 export type { ContainerQServerProps } from './components/QServer/ContainerQServer';
