+++ src/testing/tests/FinchMainContent.test.tsx
@@ -1,7 +1,9 @@
 import { render, screen } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
 import { MemoryRouter } from 'react-router';
-import FinchMainContent, { FinchMainContentProps } from '../../components/FinchMainContent';
+import FinchMainContent, {
+    FinchMainContentProps,
+} from '../../components/FinchAppLayout/FinchMainContent';
 import { RouteItem } from '../../types/navigationRouterTypes';
 
 const mockRoutes: RouteItem[] = [
